refactor: decompose 3 oversized CLI crate modules (#438)
Extract focused sibling modules from the three largest production files
in crates/cli/src/:

- tool_executor.rs (795 LOC): Extract create_schema_error (~190 LOC
  match arms) into tool_schema_errors.rs
- mcp_commands.rs (640 LOC): Extract JSON-RPC stdio server into
  mcp_serve.rs, CLI/TUI dispatch routing into mcp_dispatch.rs
- session_graph.rs (421 prod LOC): Extract persistence (load/save/path)
  into session_graph_storage.rs

Public API unchanged -- mcp_commands re-exports the dispatch functions
so all existing call sites compile without modification.
