+++ crates/cli/src/mcp_commands.rs
@@ -9,198 +9,20 @@
 //! - status: Show detailed status of a specific server
 //!
 //! Used by both CLI (claude mcp ...) and TUI (/mcp-... commands)
+//!
+//! Implementation is split across sibling modules:
+//! - `mcp_serve`: JSON-RPC 2.0 stdio server mode
+//! - `mcp_dispatch`: CLI and TUI entry-point routing
 
 use crate::plugins::mcp_proxy::McpProxy;
-use crate::schema_validator::SchemaValidator;
-use crate::tool_definitions;
-use serde::{Deserialize, Serialize};
-use std::io::{self, BufRead, Write};
 use std::sync::Arc;
 use tokio::sync::Mutex;
 
+// Re-export dispatch functions so existing call sites keep working
+pub use crate::mcp_dispatch::{handle_cli_command, handle_tui_command, parse_slash_command};
+
 /// MCP command result
 pub type McpCommandResult = Result<String, String>;
-
-/// JSON-RPC 2.0 request
-#[derive(Debug, Deserialize)]
-#[allow(dead_code)] // Fields populated by JSON deserialization
-struct JsonRpcRequest {
-    jsonrpc: String,
-    id: serde_json::Value,
-    method: String,
-    #[serde(default)]
-    params: serde_json::Value,
-}
-
-/// JSON-RPC 2.0 response
-#[derive(Debug, Serialize)]
-struct JsonRpcResponse {
-    jsonrpc: String,
-    id: serde_json::Value,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    result: Option<serde_json::Value>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    error: Option<JsonRpcError>,
-}
-
-/// JSON-RPC 2.0 error
-#[derive(Debug, Serialize)]
-struct JsonRpcError {
-    code: i32,
-    message: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    data: Option<serde_json::Value>,
-}
-
-/// Serve RustyClawd as an MCP server
-///
-/// Reads JSON-RPC 2.0 requests from stdin and writes responses to stdout.
-/// Exposes all RustyClawd tools to external MCP clients.
-async fn serve_mcp_server() -> McpCommandResult {
-    let stdin = io::stdin();
-    let mut stdout = io::stdout();
-
-    // Read requests line by line from stdin
-    for line in stdin.lock().lines() {
-        let line = line.map_err(|e| format!("Failed to read stdin: {}", e))?;
-
-        // Parse JSON-RPC request
-        let request: JsonRpcRequest = match serde_json::from_str(&line) {
-            Ok(req) => req,
-            Err(e) => {
-                // Send parse error response
-                let error_response = JsonRpcResponse {
-                    jsonrpc: "2.0".to_string(),
-                    id: serde_json::Value::Null,
-                    result: None,
-                    error: Some(JsonRpcError {
-                        code: -32700,
-                        message: "Parse error".to_string(),
-                        data: Some(serde_json::json!({ "details": e.to_string() })),
-                    }),
-                };
-
-                if let Ok(json) = serde_json::to_string(&error_response) {
-                    writeln!(stdout, "{}", json).ok();
-                    stdout.flush().ok();
-                }
-                continue;
-            }
-        };
-
-        // Handle request based on method
-        let response = match request.method.as_str() {
-            "initialize" => handle_initialize(&request),
-            "tools/list" => handle_tools_list(&request),
-            "tools/call" => handle_tools_call(&request).await,
-            _ => JsonRpcResponse {
-                jsonrpc: "2.0".to_string(),
-                id: request.id.clone(),
-                result: None,
-                error: Some(JsonRpcError {
-                    code: -32601,
-                    message: "Method not found".to_string(),
-                    data: Some(serde_json::json!({ "method": request.method })),
-                }),
-            },
-        };
-
-        // Send response
-        if let Ok(json) = serde_json::to_string(&response) {
-            writeln!(stdout, "{}", json).ok();
-            stdout.flush().ok();
-        }
-    }
-
-    Ok("MCP server stopped".to_string())
-}
-
-/// Handle initialize request
-fn handle_initialize(request: &JsonRpcRequest) -> JsonRpcResponse {
-    JsonRpcResponse {
-        jsonrpc: "2.0".to_string(),
-        id: request.id.clone(),
-        result: Some(serde_json::json!({
-            "protocolVersion": "1.0",
-            "capabilities": {
-                "tools": true
-            },
-            "serverInfo": {
-                "name": "rustyclawd",
-                "version": env!("CARGO_PKG_VERSION")
-            }
-        })),
-        error: None,
-    }
-}
-
-/// Handle tools/list request
-fn handle_tools_list(request: &JsonRpcRequest) -> JsonRpcResponse {
-    // Get all tool definitions and convert to MCP format
-    // Trust tool_definitions.rs to provide correct schemas
-    let tool_defs = tool_definitions::get_all_tool_definitions();
-    let validator = SchemaValidator::default();
-    let mut filtered_count = 0;
-
-    let tools: Vec<serde_json::Value> = tool_defs
-        .into_iter()
-        .filter_map(|tool| {
-            // Validate the input schema
-            let validation_result = validator.validate(&tool.input_schema);
-            if !validation_result.is_valid() {
-                // Log the filtered tool for debugging
-                eprintln!(
-                    "MCP serve: Filtered out tool '{}' due to incompatible schema: {}",
-                    tool.name,
-                    validation_result.error_message().unwrap_or_default()
-                );
-                filtered_count += 1;
-                return None;
-            }
-
-            Some(serde_json::json!({
-                "name": tool.name,
-                "description": tool.description,
-                "inputSchema": tool.input_schema
-            }))
-        })
-        .collect();
-
-    // Log summary if any tools were filtered
-    if filtered_count > 0 {
-        eprintln!(
-            "MCP serve: Filtered out {} tool(s) with incompatible schemas",
-            filtered_count
-        );
-    }
-
-    JsonRpcResponse {
-        jsonrpc: "2.0".to_string(),
-        id: request.id.clone(),
-        result: Some(serde_json::json!({ "tools": tools })),
-        error: None,
-    }
-}
-
-/// Handle tools/call request
-///
-/// Tool execution in MCP serve mode is not supported because it requires
-/// the complete tool_executor context (hooks, session state, permissions).
-/// Returns a proper error response per JSON-RPC 2.0 spec.
-async fn handle_tools_call(request: &JsonRpcRequest) -> JsonRpcResponse {
-    JsonRpcResponse {
-        jsonrpc: "2.0".to_string(),
-        id: request.id.clone(),
-        result: None,
-        error: Some(JsonRpcError {
-            code: -32601,
-            message: "Method not supported".to_string(),
-            data: Some(serde_json::json!({
-                "details": "Tool execution requires session context (hooks, permissions, state) which is not available in MCP serve mode. Use CLI mode for tool execution."
-            })),
-        }),
-    }
-}
 
 /// MCP command handler - shared between CLI and TUI
 pub struct McpCommandHandler {
@@ -464,177 +286,5 @@ impl McpCommandHandler {
         let mut proxy = self.proxy.lock().await;
         proxy.stop_all().await?;
         Ok("All MCP servers stopped".to_string())
-    }
-}
-
-/// CLI entry point for MCP subcommands
-pub async fn handle_cli_command(proxy: Arc<Mutex<McpProxy>>, args: &[String]) -> McpCommandResult {
-    let handler = McpCommandHandler::new(proxy);
-
-    if args.is_empty() {
-        return Err(
-            "Missing subcommand. Usage: mcp <start|stop|list|tools|prompts|status> [args]"
-                .to_string(),
-        );
-    }
-
-    let subcommand = args[0].as_str();
-
-    match subcommand {
-        "serve" => serve_mcp_server().await,
-        "start" => {
-            if args.len() < 2 {
-                return Err("Missing server ID. Usage: mcp start <server-id>".to_string());
-            }
-            handler.start(&args[1]).await
-        }
-        "stop" => {
-            if args.len() < 2 {
-                return Err("Missing server ID. Usage: mcp stop <server-id>".to_string());
-            }
-            handler.stop(&args[1]).await
-        }
-        "list" => handler.list().await,
-        "tools" => {
-            if args.len() < 2 {
-                return Err("Missing server ID. Usage: mcp tools <server-id>".to_string());
-            }
-            handler.tools(&args[1]).await
-        }
-        "prompts" => {
-            if args.len() < 2 {
-                return Err("Missing server ID. Usage: mcp prompts <server-id>".to_string());
-            }
-            handler.prompts(&args[1]).await
-        }
-        "status" => {
-            if args.len() < 2 {
-                return Err("Missing server ID. Usage: mcp status <server-id>".to_string());
-            }
-            handler.status(&args[1]).await
-        }
-        "stop-all" => handler.stop_all().await,
-        _ => Err(format!(
-            "Unknown subcommand: '{}'. Available: serve, start, stop, list, tools, prompts, status",
-            subcommand
-        )),
-    }
-}
-
-/// Parse slash command for TUI
-/// Examples: "/mcp-list", "/mcp-start filesystem", "/mcp-tools filesystem"
-pub fn parse_slash_command(input: &str) -> Option<(String, Vec<String>)> {
-    let input = input.trim();
-
-    if !input.starts_with("/mcp-") {
-        return None;
-    }
-
-    let parts: Vec<&str> = input[1..].split_whitespace().collect();
-    if parts.is_empty() {
-        return None;
-    }
-
-    // Convert /mcp-start -> start
-    let command = parts[0].strip_prefix("mcp-")?;
-
-    // Rest are arguments
-    let args: Vec<String> = parts[1..].iter().map(|s| s.to_string()).collect();
-
-    Some((command.to_string(), args))
-}
-
-/// TUI entry point for /mcp-* commands
-pub async fn handle_tui_command(
-    proxy: Arc<Mutex<McpProxy>>,
-    command: &str,
-    args: Vec<String>,
-) -> McpCommandResult {
-    let handler = McpCommandHandler::new(proxy);
-
-    match command {
-        "start" => {
-            if args.is_empty() {
-                return Err("Missing server ID. Usage: /mcp-start <server-id>".to_string());
-            }
-            handler.start(&args[0]).await
-        }
-        "stop" => {
-            if args.is_empty() {
-                return Err("Missing server ID. Usage: /mcp-stop <server-id>".to_string());
-            }
-            handler.stop(&args[0]).await
-        }
-        "list" => handler.list().await,
-        "tools" => {
-            if args.is_empty() {
-                return Err("Missing server ID. Usage: /mcp-tools <server-id>".to_string());
-            }
-            handler.tools(&args[0]).await
-        }
-        "prompts" => {
-            if args.is_empty() {
-                return Err("Missing server ID. Usage: /mcp-prompts <server-id>".to_string());
-            }
-            handler.prompts(&args[0]).await
-        }
-        "status" => {
-            if args.is_empty() {
-                return Err("Missing server ID. Usage: /mcp-status <server-id>".to_string());
-            }
-            handler.status(&args[0]).await
-        }
-        _ => Err(format!(
-            "Unknown MCP command: '{}'. Available: start, stop, list, tools, prompts, status",
-            command
-        )),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_parse_slash_command_list() {
-        let result = parse_slash_command("/mcp-list");
-        assert!(result.is_some());
-        let (cmd, args) = result.unwrap();
-        assert_eq!(cmd, "list");
-        assert!(args.is_empty());
-    }
-
-    #[test]
-    fn test_parse_slash_command_start() {
-        let result = parse_slash_command("/mcp-start filesystem");
-        assert!(result.is_some());
-        let (cmd, args) = result.unwrap();
-        assert_eq!(cmd, "start");
-        assert_eq!(args, vec!["filesystem"]);
-    }
-
-    #[test]
-    fn test_parse_slash_command_tools() {
-        let result = parse_slash_command("/mcp-tools filesystem");
-        assert!(result.is_some());
-        let (cmd, args) = result.unwrap();
-        assert_eq!(cmd, "tools");
-        assert_eq!(args, vec!["filesystem"]);
-    }
-
-    #[test]
-    fn test_parse_slash_command_prompts() {
-        let result = parse_slash_command("/mcp-prompts filesystem");
-        assert!(result.is_some());
-        let (cmd, args) = result.unwrap();
-        assert_eq!(cmd, "prompts");
-        assert_eq!(args, vec!["filesystem"]);
-    }
-
-    #[test]
-    fn test_parse_slash_command_invalid() {
-        assert!(parse_slash_command("/help").is_none());
-        assert!(parse_slash_command("mcp-list").is_none());
-        assert!(parse_slash_command("/mcplist").is_none());
     }
 }
